Divide RequiresCapability file (#2508)
Divide RequiresCapability file into several testing files to have a better debugging experience
