+++ test/ILLink.RoslynAnalyzer.Tests/RequiresCapabilityTests.cs
@@ -9,9 +9,21 @@ namespace ILLink.RoslynAnalyzer.Tests
 		protected override string TestSuiteName => "RequiresCapability";
 
 		[Fact]
-		public Task RequiresCapability ()
+		public Task BasicRequires ()
 		{
-			return RunTest (nameof (RequiresCapability));
+			return RunTest (nameof (BasicRequires));
+		}
+
+		[Fact]
+		public Task RequiresAccessedThrough ()
+		{
+			return RunTest (nameof (RequiresAccessedThrough));
+		}
+
+		[Fact]
+		public Task RequiresAttributeMismatch ()
+		{
+			return RunTest (nameof (RequiresAttributeMismatch));
 		}
 
 		[Fact]
@@ -33,9 +45,57 @@ namespace ILLink.RoslynAnalyzer.Tests
 		}
 
 		[Fact]
+		public Task RequiresOnAttribute ()
+		{
+			return RunTest (nameof (RequiresOnAttribute));
+		}
+
+		[Fact]
 		public Task RequiresOnAttributeCtor ()
 		{
 			return RunTest (nameof (RequiresOnAttributeCtor));
+		}
+
+		[Fact]
+		public Task RequiresOnClass ()
+		{
+			return RunTest (nameof (RequiresOnClass));
+		}
+
+		[Fact]
+		public Task RequiresOnStaticConstructor ()
+		{
+			return RunTest (nameof (RequiresOnStaticConstructor));
+		}
+
+		[Fact]
+		public Task RequiresOnVirtualsAndInterfaces ()
+		{
+			return RunTest (nameof (RequiresOnVirtualsAndInterfaces));
+		}
+
+		[Fact]
+		public Task RequiresViaDataflow ()
+		{
+			return RunTest (nameof (RequiresViaDataflow));
+		}
+
+		[Fact]
+		public Task RequiresViaXml ()
+		{
+			return RunTest (nameof (RequiresViaXml));
+		}
+
+		[Fact]
+		public Task RequiresWithCopyAssembly ()
+		{
+			return RunTest (nameof (RequiresWithCopyAssembly));
+		}
+
+		[Fact]
+		public Task SuppressRequires ()
+		{
+			return RunTest (nameof (SuppressRequires));
 		}
 	}
 }
