+++ test/Mono.Linker.Tests.Cases/RequiresCapability/RequiresAccessedThrough.cs
@@ -1,0 +1,171 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Threading.Tasks;
+using Mono.Linker.Tests.Cases.Expectations.Assertions;
+
+namespace Mono.Linker.Tests.Cases.RequiresCapability
+{
+	[SkipKeptItemsValidation]
+	[ExpectedNoWarnings]
+	class RequiresAccessedThrough
+	{
+		public static void Main ()
+		{
+			TestRequiresOnlyThroughReflection ();
+			AccessedThroughReflectionOnGenericType<TestType>.Test ();
+			AccessThroughSpecialAttribute.Test ();
+			AccessThroughPInvoke.Test ();
+			AccessThroughNewConstraint.Test ();
+			AccessThroughNewConstraint.TestNewConstraintOnTypeParameter ();
+			AccessThroughNewConstraint.TestNewConstraintOnTypeParameterOfStaticType ();
+			AccessThroughLdToken.Test ();
+		}
+
+		class TestType { }
+
+		[RequiresUnreferencedCode ("Message for --RequiresOnlyThroughReflection--")]
+		static void RequiresOnlyThroughReflection ()
+		{
+		}
+
+		[ExpectedWarning ("IL2026", "--RequiresOnlyThroughReflection--", ProducedBy = ProducedBy.Trimmer)]
+		static void TestRequiresOnlyThroughReflection ()
+		{
+			typeof (RequiresAccessedThrough)
+				.GetMethod (nameof (RequiresOnlyThroughReflection), System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic)
+				.Invoke (null, new object[0]);
+		}
+
+		class AccessedThroughReflectionOnGenericType<T>
+		{
+			[RequiresUnreferencedCode ("Message for --GenericType.RequiresOnlyThroughReflection--")]
+			public static void RequiresOnlyThroughReflection ()
+			{
+			}
+
+			[ExpectedWarning ("IL2026", "--GenericType.RequiresOnlyThroughReflection--", ProducedBy = ProducedBy.Trimmer)]
+			public static void Test ()
+			{
+				typeof (AccessedThroughReflectionOnGenericType<T>)
+					.GetMethod (nameof (RequiresOnlyThroughReflection))
+					.Invoke (null, new object[0]);
+			}
+		}
+
+		class AccessThroughSpecialAttribute
+		{
+			// https://github.com/dotnet/linker/issues/1873
+			// [ExpectedWarning ("IL2026", "--DebuggerProxyType.Method--")]
+			[DebuggerDisplay ("Some{*}value")]
+			class TypeWithDebuggerDisplay
+			{
+				[RequiresUnreferencedCode ("Message for --DebuggerProxyType.Method--")]
+				public void Method ()
+				{
+				}
+			}
+
+			public static void Test ()
+			{
+				var _ = new TypeWithDebuggerDisplay ();
+			}
+		}
+
+		class AccessThroughPInvoke
+		{
+			class PInvokeReturnType
+			{
+				[RequiresUnreferencedCode ("Message for --PInvokeReturnType.ctor--")]
+				public PInvokeReturnType () { }
+			}
+
+			// https://github.com/mono/linker/issues/2116
+			[ExpectedWarning ("IL2026", "--PInvokeReturnType.ctor--", ProducedBy = ProducedBy.Trimmer)]
+			[DllImport ("nonexistent")]
+			static extern PInvokeReturnType PInvokeReturnsType ();
+
+			// Analyzer doesn't support IL2050 yet
+			[ExpectedWarning ("IL2050", ProducedBy = ProducedBy.Trimmer)]
+			public static void Test ()
+			{
+				PInvokeReturnsType ();
+			}
+		}
+
+		class AccessThroughNewConstraint
+		{
+			class NewConstraintTestType
+			{
+				[RequiresUnreferencedCode ("Message for --NewConstraintTestType.ctor--")]
+				[RequiresAssemblyFiles ("Message for --NewConstraintTestType.ctor--")]
+				[RequiresDynamicCode ("Message for --NewConstraintTestType.ctor--")]
+				public NewConstraintTestType () { }
+			}
+
+			static void GenericMethod<T> () where T : new() { }
+
+			[ExpectedWarning ("IL2026", "--NewConstraintTestType.ctor--")]
+			[ExpectedWarning ("IL3002", "--NewConstraintTestType.ctor--", ProducedBy = ProducedBy.Analyzer)]
+			[ExpectedWarning ("IL3050", "--NewConstraintTestType.ctor--", ProducedBy = ProducedBy.Analyzer)]
+			public static void Test ()
+			{
+				GenericMethod<NewConstraintTestType> ();
+			}
+
+			static class NewConstraintOnTypeParameterOfStaticType<T> where T : new()
+			{
+				public static void DoNothing () { }
+			}
+
+			class NewConstaintOnTypeParameter<T> where T : new()
+			{
+			}
+
+			[ExpectedWarning ("IL2026", "--NewConstraintTestType.ctor--")]
+			[ExpectedWarning ("IL3002", "--NewConstraintTestType.ctor--", ProducedBy = ProducedBy.Analyzer)]
+			[ExpectedWarning ("IL3050", "--NewConstraintTestType.ctor--", ProducedBy = ProducedBy.Analyzer)]
+			public static void TestNewConstraintOnTypeParameter ()
+			{
+				_ = new NewConstaintOnTypeParameter<NewConstraintTestType> ();
+			}
+
+			[ExpectedWarning ("IL2026", "--NewConstraintTestType.ctor--")]
+			[ExpectedWarning ("IL3002", "--NewConstraintTestType.ctor--", ProducedBy = ProducedBy.Analyzer)]
+			[ExpectedWarning ("IL3050", "--NewConstraintTestType.ctor--", ProducedBy = ProducedBy.Analyzer)]
+			public static void TestNewConstraintOnTypeParameterOfStaticType ()
+			{
+				NewConstraintOnTypeParameterOfStaticType<NewConstraintTestType>.DoNothing ();
+			}
+		}
+
+		class AccessThroughLdToken
+		{
+			static bool PropertyWithLdToken {
+				[RequiresUnreferencedCode ("Message for --PropertyWithLdToken.get--")]
+				[RequiresAssemblyFiles ("Message for --PropertyWithLdToken.get--")]
+				[RequiresDynamicCode ("Message for --PropertyWithLdToken.get--")]
+				get {
+					return false;
+				}
+			}
+
+			[ExpectedWarning ("IL2026", "--PropertyWithLdToken.get--")]
+			[ExpectedWarning ("IL2026", "--PropertyWithLdToken.get--", ProducedBy = ProducedBy.Trimmer)]
+			[ExpectedWarning ("IL3002", "--PropertyWithLdToken.get--", ProducedBy = ProducedBy.Analyzer)]
+			[ExpectedWarning ("IL3050", "--PropertyWithLdToken.get--", ProducedBy = ProducedBy.Analyzer)]
+			public static void Test ()
+			{
+				Expression<Func<bool>> getter = () => PropertyWithLdToken;
+			}
+		}
+	}
+}
