+++ test/Mono.Linker.Tests.Cases/RequiresCapability/RequiresViaXml.attributes.xml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <linker>
   <assembly fullname="test, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null">
-    <type fullname="Mono.Linker.Tests.Cases.RequiresCapability.RequiresCapability">
+    <type fullname="Mono.Linker.Tests.Cases.RequiresCapability.RequiresViaXml">
       <method name="MethodWithDuplicateRequiresAttribute">
         <attribute fullname="System.Diagnostics.CodeAnalysis.RequiresUnreferencedCodeAttribute">
           <argument>Message for MethodWithDuplicateRequiresAttribute from link attributes XML</argument>
