+++ test/Mono.Linker.Tests.Cases/RequiresCapability/RequiresViaXml.descriptor.xml
@@ -1,9 +1,9 @@
 ﻿<linker>
   <assembly fullname="test, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null">
-    <type fullname="Mono.Linker.Tests.Cases.RequiresCapability.RequiresCapability">
+    <type fullname="Mono.Linker.Tests.Cases.RequiresCapability.RequiresViaXml">
       <method name="RequiresOnlyViaDescriptor" />
     </type>
-    <type fullname="Mono.Linker.Tests.Cases.RequiresCapability.RequiresCapability/RequiresOnFieldOnlyViaDescriptor">
+    <type fullname="Mono.Linker.Tests.Cases.RequiresCapability.RequiresViaXml/RequiresOnFieldOnlyViaDescriptor">
       <field name="Field" />
     </type>
   </assembly>
